Adding close button to popup.
