+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/detail/RestaurantDetail.kt
@@ -7,12 +7,8 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowForward
-import androidx.compose.material.icons.filled.Phone
-import androidx.compose.material.icons.filled.Place
-import androidx.compose.material.icons.filled.Send
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
@@ -24,18 +20,9 @@ import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Popup
-import com.google.android.gms.maps.model.CameraPosition
-import com.google.android.gms.maps.model.LatLng
-import com.google.maps.android.compose.GoogleMap
-import com.google.maps.android.compose.Marker
-import com.google.maps.android.compose.MarkerState
-import com.google.maps.android.compose.rememberCameraPositionState
 import com.jordangellatly.hermosahappyhour.R
-import com.jordangellatly.hermosahappyhour.model.Location
 import com.jordangellatly.hermosahappyhour.model.Restaurant
 import com.jordangellatly.hermosahappyhour.model.RestaurantRepo
-import com.jordangellatly.hermosahappyhour.ui.components.HappyHourDivider
 import com.jordangellatly.hermosahappyhour.ui.home.getCurrentDateTime
 import com.jordangellatly.hermosahappyhour.ui.home.toString
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
@@ -260,148 +247,6 @@ private fun Specials(restaurant: Restaurant?) {
             text = "N/A",
             modifier = Modifier.padding(start = 8.dp, end = 8.dp)
         )
-    }
-}
-
-@Composable
-private fun GeneralInfo(restaurant: Restaurant?) {
-    Column(
-        modifier = Modifier.padding(8.dp)
-    ) {
-        val date = getCurrentDateTime()
-        val dateInString = date.toString("EEEE").uppercase()
-        val dailyInfo =
-            restaurant?.hoursAndSpecials?.find { it.dayOfWeek.toString() == dateInString }
-        var popupControl by remember { mutableStateOf(false) }
-        Text(
-            text = "Info",
-            style = MaterialTheme.typography.h4,
-            modifier = Modifier.padding(8.dp)
-        )
-        Row(
-            horizontalArrangement = Arrangement.SpaceBetween,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Column {
-                Text(
-                    text = "Hours",
-                    style = MaterialTheme.typography.h6,
-                    modifier = Modifier.padding(start = 8.dp, end = 8.dp)
-                )
-                Text(
-                    text = dailyInfo?.businessHours.toString(),
-                    modifier = Modifier.padding(start = 8.dp, end = 8.dp, bottom = 8.dp)
-                )
-            }
-            IconButton(onClick = { popupControl = true }) {
-                Icon(
-                    imageVector = Icons.Filled.ArrowForward,
-                    contentDescription = "hours"
-                )
-            }
-        }
-        HappyHourDivider()
-
-        Row(
-            horizontalArrangement = Arrangement.SpaceBetween,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Column {
-                Text(
-                    text = "Website",
-                    style = MaterialTheme.typography.h6,
-                    modifier = Modifier.padding(start = 8.dp, end = 8.dp)
-                )
-                Text(
-                    text = restaurant?.website.toString(),
-                    modifier = Modifier.padding(start = 8.dp, end = 8.dp, bottom = 8.dp)
-                )
-            }
-            IconButton(onClick = { /*TODO*/ }) {
-                Icon(
-                    imageVector = Icons.Filled.Send,
-                    contentDescription = "website"
-                )
-            }
-        }
-        HappyHourDivider()
-
-        Row(
-            horizontalArrangement = Arrangement.SpaceBetween,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Column {
-                Text(
-                    text = "Call",
-                    style = MaterialTheme.typography.h6,
-                    modifier = Modifier.padding(start = 8.dp, end = 8.dp)
-                )
-                Text(
-                    text = restaurant?.phoneNumber ?: "",
-                    modifier = Modifier.padding(start = 8.dp, end = 8.dp, bottom = 8.dp)
-                )
-            }
-            IconButton(onClick = { /*TODO*/ }) {
-                Icon(
-                    imageVector = Icons.Filled.Phone,
-                    contentDescription = "phone"
-                )
-            }
-        }
-        HappyHourDivider()
-
-        Row(
-            horizontalArrangement = Arrangement.SpaceBetween,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Column {
-                Text(
-                    text = "Get Directions",
-                    style = MaterialTheme.typography.h6,
-                    modifier = Modifier.padding(start = 8.dp, end = 8.dp)
-                )
-                Text(
-                    text = restaurant?.address?.line1 ?: "",
-                    modifier = Modifier.padding(start = 8.dp, end = 8.dp)
-                )
-                Text(
-                    text = restaurant?.address?.line2 ?: "",
-                    modifier = Modifier.padding(start = 8.dp, end = 8.dp, bottom = 8.dp)
-                )
-            }
-            IconButton(onClick = { /*TODO*/ }) {
-                Icon(
-                    imageVector = Icons.Filled.Place,
-                    contentDescription = "get directions"
-                )
-            }
-        }
-        HappyHourDivider()
-
-        val location = restaurant?.location ?: Location(0.0, 0.0)
-        val latlng = LatLng(location.latitude, location.longitude)
-        val cameraPositionState = rememberCameraPositionState {
-            position = CameraPosition.fromLatLngZoom(latlng, 16f)
-        }
-        GoogleMap(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(200.dp),
-            cameraPositionState = cameraPositionState
-        ) {
-            Marker(
-                state = MarkerState(position = latlng),
-                title = restaurant?.name
-            )
-        }
-
-        if (popupControl) {
-            Popup(
-                onDismissRequest = { popupControl = false }
-            ) {
-                RestaurantHours(restaurant = restaurant)
-            }
-        }
     }
 }
