+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/detail/RestaurantHours.kt
@@ -3,9 +3,15 @@ package com.jordangellatly.hermosahappyhour.ui.detail
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
@@ -15,9 +21,13 @@ import com.jordangellatly.hermosahappyhour.model.Restaurant
 import com.jordangellatly.hermosahappyhour.model.tower12RestaurantData
 import com.jordangellatly.hermosahappyhour.ui.components.HappyHourDivider
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
+import com.jordangellatly.hermosahappyhour.ui.theme.Neutral8
 
 @Composable
-fun RestaurantHours(restaurant: Restaurant?) {
+fun RestaurantHours(
+    restaurant: Restaurant?,
+    onClick: () -> Unit
+) {
     Column(
         modifier = Modifier
             .background(Color.White)
@@ -27,11 +37,21 @@ fun RestaurantHours(restaurant: Restaurant?) {
             )
             .padding(8.dp)
     ) {
-        Text(
-            text = "Hours",
-            style = MaterialTheme.typography.h4,
-            modifier = Modifier.padding(16.dp)
-        )
+        Row(
+            horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            Text(
+                text = "Hours",
+                style = MaterialTheme.typography.h4,
+                modifier = Modifier.padding(16.dp)
+            )
+            Close(
+                onClick = onClick
+            )
+        }
+
         DayOfWeek.values().forEach { dayOfWeek ->
             val dailyInfo = restaurant?.hoursAndSpecials?.find { it.dayOfWeek == dayOfWeek }
             val businessHours = dailyInfo?.businessHours
@@ -59,12 +79,34 @@ fun RestaurantHours(restaurant: Restaurant?) {
     }
 }
 
+@Composable
+private fun Close(onClick: () -> Unit) {
+    IconButton(
+        onClick = onClick,
+        modifier = Modifier
+            .statusBarsPadding()
+            .padding(horizontal = 16.dp, vertical = 10.dp)
+            .size(36.dp)
+            .background(
+                color = Neutral8.copy(alpha = 0.32f),
+                shape = CircleShape
+            )
+    ) {
+        Icon(
+            imageVector = Icons.Filled.Close,
+            tint = HermosaHappyHourTheme.colors.iconInteractive,
+            contentDescription = "close"
+        )
+    }
+}
+
 @Preview(showBackground = true)
 @Composable
 private fun RestaurantHoursPreview() {
     HermosaHappyHourTheme {
         RestaurantHours(
-            restaurant = tower12RestaurantData
+            restaurant = tower12RestaurantData,
+            onClick = {}
         )
     }
 }
